// lib/src/client/types/blue/moji/collection/putItem.ts
/**
 * GENERATED CODE - DO NOT MODIFY
 */
import { HeadersMap, XRPCError } from "@atproto/xrpc";
import { type ValidationResult, BlobRef } from "@atproto/lexicon";
import { CID } from "multiformats/cid";
import { validate as _validate } from "../../../../lexicons";
import { type $Typed, is$typed as _is$typed, type OmitKey } from "../../../../util";
import type * as BlueMojiCollectionItem from "./item.js";

const is$typed = _is$typed,
  validate = _validate;
const id = "blue.moji.collection.putItem";

export type QueryParams = {};

export interface InputSchema {
  /** The handle or DID of the repo (aka, current account). */
  repo: string;
  /** Can be set to 'false' to skip Lexicon schema validation of record data. */
  validate?: boolean;
  item: BlueMojiCollectionItem.ItemView;
}

export interface OutputSchema {
  uri: string;
}

export interface CallOptions {
  signal?: AbortSignal;
  headers?: HeadersMap;
  qp?: QueryParams;
  encoding?: "application/json";
}

export interface Response {
  success: boolean;
  headers: HeadersMap;
  data: OutputSchema;
}

export function toKnownErr(e: any) {
  return e;
}